requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
weasyprint
pdfplumber
matplotlib
python-dotenv
python-multipart
